s3gk expose internal API in addition to CLI tool.
* artifactory client utils;
* meta pointer related factilies;
* package version auto computed (based on already published package  version);
