+++ package.json
@@ -4,6 +4,9 @@
     "build:prod": "wp --config webpack.config.js --mode production",
     "lint": "eslint --ext .ts,.tsx src/"
   },
+  "exports": {
+    ".": "./api.js"
+  },
   "private": true,
   "workspaces": [],
   "devDependencies": {
@@ -21,6 +24,7 @@
     "file-loader": "^6.2.0",
     "generate-json-webpack-plugin": "^2.0.0",
     "minimatch": "3.0.4",
+    "semver": "^7.3.5",
     "ts-loader": "^8.0.7",
     "ts-node-dev": "^1.0.0-pre.63",
     "typescript": "^4.0.5",
