+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export { MetaPointer, readMetaPointerFromFile } from './metapointer';
+export * from './ArtifactoryClient';
