+++ tsconfig.json
@@ -14,6 +14,7 @@
 			"es5"
 		],
 		"module": "ES6",
+		"declaration": false,
 		"moduleResolution": "Node",
 		"noImplicitAny": true,
 		"sourceMap": true,
